전체적으로 수정

+++ powershell/W-27.ps1
@@ -4,7 +4,7 @@ $json = @{
     Code = "W-27"
     RiskLevel = "높음"
     DiagnosticItem = "비밀번호 저장을 위한 복호화 가능한 암호화 사용"
-    DiagnosticResult = "양호" # 기본 상태를 '양호'로 가정
+    DiagnosticResult = "양호"  # 기본 상태를 '양호'로 가정
     CurrentStatus = @()
     Recommendation = "비밀번호 저장을 위한 복호화 가능한 암호화 사용"
 }
@@ -19,32 +19,36 @@ if (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdenti
 $computerName = $env:COMPUTERNAME
 $rawDir = "C:\Window_${computerName}_raw"
 $resultDir = "C:\Window_${computerName}_result"
-Remove-Item -Path $rawDir, $resultDir -Recurse -ErrorAction SilentlyContinue
-New-Item -Path $rawDir, $resultDir -ItemType Directory | Out-Null
+Remove-Item -Path $rawDir, $resultDir -Recurse -ErrorAction SilentlyContinue -Force
+New-Item -Path $rawDir, $resultDir -ItemType Directory -Force | Out-Null
 secedit /export /cfg "$rawDir\Local_Security_Policy.txt"
 $null = New-Item -Path "$rawDir\compare.txt" -ItemType File
 Set-Location -Path $rawDir
-(Get-Location).Path | Out-File "$rawDir\install_path.txt"
+[System.IO.File]::WriteAllText("$rawDir\install_path.txt", (Get-Location).Path)
 systeminfo | Out-File "$rawDir\systeminfo.txt"
 
-# Analyze IIS configuration
-Get-Content "$env:WinDir\System32\Inetsrv\Config\applicationHost.Config" | Out-File "$rawDir\iis_setting.txt"
-Select-String -Path "$rawDir\iis_setting.txt" -Pattern "physicalPath|bindingInformation" | Out-File "$rawDir\iis_path1.txt"
+# IIS 설정 분석
+$applicationHostConfigPath = "$env:WinDir\System32\Inetsrv\Config\applicationHost.Config"
+if (Test-Path $applicationHostConfigPath) {
+    Get-Content $applicationHostConfigPath | Out-File "$rawDir\iis_setting.txt"
+    Select-String -Path "$rawDir\iis_setting.txt" -Pattern "physicalPath|bindingInformation" | Out-File "$rawDir\iis_path1.txt"
+} else {
+    Write-Output "IIS configuration file not found."
+}
 
 # IISADMIN 서비스 계정 검사
 $serviceStatus = Get-Service W3SVC -ErrorAction SilentlyContinue
 if ($serviceStatus.Status -eq 'Running') {
     $iisAdminReg = Get-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\IISADMIN" -Name "ObjectName" -ErrorAction SilentlyContinue
-    if ($iisAdminReg.ObjectName -ne "LocalSystem") {
+    if ($iisAdminReg -and $iisAdminReg.ObjectName -ne "LocalSystem") {
         $json.CurrentStatus += "IISADMIN 서비스가 LocalSystem 계정에서 실행되지 않고 있습니다. 특별한 조치가 필요하지 않습니다."
-    } else {
+    } elseif ($iisAdminReg) {
         $json.CurrentStatus += "IISADMIN 서비스가 LocalSystem 계정에서 실행되고 있습니다. 권장되지 않습니다."
     }
 } else {
     $json.CurrentStatus += "월드 와이드 웹 퍼블리싱 서비스가 실행되지 않고 있습니다. IIS 관련 보안 구성 검토가 필요 없습니다."
 }
 
-
-# Save the JSON results to a file
+# JSON 결과를 파일에 저장
 $jsonFilePath = "$resultDir\W-27.json"
 $json | ConvertTo-Json -Depth 3 | Out-File -FilePath $jsonFilePath
