+++ powershell/W-28.ps1
@@ -9,28 +9,33 @@ $json = @{
     Recommendation = "비밀번호 저장에 복호화 가능한 암호화 사용을 피하세요"
 }
 
-# Request administrator privileges
+# 관리자 권한 요청
 if (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)) {
     Start-Process PowerShell.exe -ArgumentList "-NoProfile", "-ExecutionPolicy Bypass", "-File `"$PSCommandPath`"", "-Verb RunAs"
     exit
 }
 
-# Setup environment and directories
+# 환경 및 디렉터리 설정
 $computerName = $env:COMPUTERNAME
 $rawDir = "C:\Window_${computerName}_raw"
 $resultDir = "C:\Window_${computerName}_result"
 Remove-Item -Path $rawDir, $resultDir -Recurse -ErrorAction SilentlyContinue
-New-Item -Path $rawDir, $resultDir -ItemType Directory
+New-Item -Path $rawDir, $resultDir -ItemType Directory | Out-Null
 secedit /export /cfg "$rawDir\Local_Security_Policy.txt"
 $null = New-Item -Path "$rawDir\compare.txt" -ItemType File
 Set-Location -Path $rawDir
 [System.IO.File]::WriteAllText("$rawDir\install_path.txt", (Get-Location).Path)
 systeminfo | Out-File "$rawDir\systeminfo.txt"
 
-# Analyze IIS configuration
-$applicationHostConfig = Get-Content "$env:WinDir\System32\Inetsrv\Config\applicationHost.Config"
-$applicationHostConfig | Out-File "$rawDir\iis_setting.txt"
-$applicationHostConfig | Select-String -Pattern "physicalPath|bindingInformation" | Out-File "$rawDir\iis_path1.txt"
+# IIS 설정 분석
+$applicationHostConfigPath = "$env:WinDir\System32\Inetsrv\Config\applicationHost.Config"
+if (Test-Path $applicationHostConfigPath) {
+    $applicationHostConfig = Get-Content $applicationHostConfigPath
+    $applicationHostConfig | Out-File "$rawDir\iis_setting.txt"
+    $applicationHostConfig | Select-String -Pattern "physicalPath|bindingInformation" | Out-File "$rawDir\iis_path1.txt"
+} else {
+    "$applicationHostConfigPath 경로를 찾을 수 없습니다." | Out-File "$rawDir\iis_path1.txt"
+}
 
 # IIS 중요 경로에서 단축 파일 검사
 $serviceStatus = Get-Service W3SVC -ErrorAction SilentlyContinue
@@ -57,6 +62,6 @@ if ($serviceStatus.Status -eq 'Running') {
     $json.CurrentStatus += "World Wide Web Publishing Service가 실행되지 않고 있습니다, 단축 파일 검사가 필요 없습니다."
 }
 
-# Save the JSON results to a file
+# JSON 결과를 파일에 저장
 $jsonFilePath = "$resultDir\W-28.json"
 $json | ConvertTo-Json -Depth 3 | Out-File -FilePath $jsonFilePath
